+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Low to Low now only bakes itself and ignores the rest of the objects
 ### Changed
 - The preview cage is now unselectable and selection and active object is kept when showing/hiding it
+- The low to low setting is now set per baker and not per bake_group
 
 ## [0.0.3](https://gitlab.com/AquaticNightmare/ez_baker/-/releases/0_0_3)
 ### Fixed